Move business object models to the separate folder
