+++ src/ggrc-client/js/models/business-models/index.js
@@ -1,0 +1,43 @@
+/*
+  Copyright (C) 2018 Google Inc.
+  Licensed under http:/www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
+*/
+
+import './access-group';
+import './assessment';
+import './assessment-template';
+import './audit';
+import './clause';
+import './contract';
+import './control';
+import './cycle-task-group-object-task';
+import './cycle-task-group';
+import './cycle';
+import './data-asset';
+import './directive';
+import './document';
+import './evidence';
+import './facility';
+import './issue';
+import './market';
+import './metric';
+import './objective';
+import './org-group';
+import './person';
+import './policy';
+import './process';
+import './product';
+import './program';
+import './project';
+import './regulation';
+import './risk';
+import './risk-assessment';
+import './section';
+import './standard';
+import './system';
+import './system-or-process';
+import './task-group';
+import './task-group-task';
+import './threat';
+import './vendor';
+import './workflow';
