+++ src/ggrc-client/js/models/business-models/metric.js
@@ -1,0 +1,70 @@
+/*
+    Copyright (C) 2018 Google Inc.
+    Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
+*/
+
+export default can.Model.Cacheable('CMS.Models.Metric', {
+  root_object: 'metric',
+  root_collection: 'metrics',
+  category: 'business',
+  findAll: '/api/metrics',
+  findOne: '/api/metrics/{id}',
+  create: 'POST /api/metrics',
+  update: 'PUT /api/metrics/{id}',
+  destroy: 'DELETE /api/metrics/{id}',
+  mixins: [
+    'unique_title',
+    'ca_update',
+    'timeboxed',
+    'accessControlList',
+    'base-notifications',
+  ],
+  attributes: {
+    context: 'CMS.Models.Context.stub',
+    modified_by: 'CMS.Models.Person.stub',
+    object_people: 'CMS.Models.ObjectPerson.stubs',
+    people: 'CMS.Models.Person.stubs',
+    objectives: 'CMS.Models.Objective.stubs',
+    controls: 'CMS.Models.Control.stubs',
+    sections: 'CMS.Models.get_stubs',
+  },
+  tree_view_options: {
+    attr_view: GGRC.mustache_path + '/base_objects/tree-item-attr.mustache',
+    attr_list: can.Model.Cacheable.attr_list.concat([
+      {attr_title: 'Effective Date', attr_name: 'start_date'},
+      {attr_title: 'Last Deprecated Date', attr_name: 'end_date'},
+      {attr_title: 'Reference URL', attr_name: 'reference_url'},
+      {
+        attr_title: 'Description',
+        attr_name: 'description',
+        disable_sorting: true,
+      }, {
+        attr_title: 'Notes',
+        attr_name: 'notes',
+        disable_sorting: true,
+      }, {
+        attr_title: 'Assessment Procedure',
+        attr_name: 'test_plan',
+        disable_sorting: true,
+      },
+    ]),
+    add_item_view: GGRC.mustache_path + '/base_objects/tree_add_item.mustache',
+    link_buttons: true,
+    display_attr_names: ['title', 'status', 'updated_at'],
+  },
+  is_custom_attributable: true,
+  isRoleable: true,
+  defaults: {
+    title: '',
+    url: '',
+    status: 'Draft',
+  },
+  sub_tree_view_options: {
+    default_filter: ['Product'],
+  },
+  statuses: ['Draft', 'Deprecated', 'Active'],
+  init: function () {
+    this.validateNonBlank('title');
+    this._super(...arguments);
+  },
+}, {});
